Add concept_ref/pipe_ref aliases, fix concept_cmd alias bug, and reorganize spec tests
Add "concept_ref" and "pipe_ref" as accepted aliases in parse_concept_spec and parse_pipe_spec for better AI agent compatibility. Fix _parse_concept_spec_from_json which incorrectly listed "concept_code" as an alias instead of "the_concept_code". Reorganize concept spec tests from cli/ to builder/operations/, add dedicated test modules for parse_concept_spec and shared test_data, and expand pipe spec test coverage.
